Update issue-configure-call-queue.md
@roykuntz , @CarolynRowe , @simonxjx 

Updated this to match current scenario. We do support hybrid with Teams CQ/AA. Service is now Teams not SfB. If the hosting provider for resources.lync.com is not configured that transfers from agent to another in tenant agent will fail. 

This document needs to move to the Teams branch. 

   
SIPPROXY_E_EPROUTING_MSG_INVALID_EXT_SPLIT_DOMAIN